pyproject.toml
[project]
name = "likwid-roofline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
